chore: Added a few of the suggestions on the rustyline code
