+++ internal/cmd/headless_repl/src/processes/mod.rs
@@ -1,14 +1,12 @@
+pub use coordinator_impl::run;
 pub use invoke_go::{read_json_rpc, start_go};
-pub use lsp_invoke::{
-    add_headers, formulate_request, join_imports, start_lsp, LSPRequestType, LSP_Error,
-};
+pub use lsp_invoke::{add_headers, formulate_request, start_lsp, LSPRequestType, LSP_Error};
 pub use process_completion::process_completions_response;
-pub use CoordinatorImpl::run;
-mod CoordinatorImpl;
-pub mod FluxServerImpl;
-mod LSPServerImpl;
+mod coordinator_impl;
+pub mod flux_server_impl;
 pub mod invoke_go;
 pub mod lsp_invoke;
+mod lsp_server_impl;
 pub mod process_completion;
 
 //run once tag
