--- internal/cmd/headless_repl/src/tests/_test_repl.go
@@ -1,42 +1,0 @@
-package tests
-
-import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"net"
-	"net/rpc"
-	"reflect"
-	"strings"
-	"testing"
-)
-
-
-type Args struct {
-	A, B int
-}
-
-type Reply struct {
-	C int
-}
-
-type Arith int
-
-type ArithAddResp struct {
-	Id     any   `json:"id"`
-	Result Reply `json:"result"`
-	Error  any   `json:"error"`
-}
-
-func (t *Arith) Add(args *Args, reply *Reply) error {
-	reply.C = args.A + args.B
-	return nil
-}
-
-func (t *Arith) Mul(args *Args, reply *Reply) error {
-	reply.C = args.A * args.B
-	return nil
-}
-
