--- internal/cmd/headless_repl/src/tests/go.mod
@@ -1,3 +1,0 @@
-module tests
-
-go 1.18
